+++ HISTORY.rst
@@ -21,6 +21,8 @@ latest (unreleased)
 
 **Features and Improvements**
 
+* ALCYN-1982: Do not export to the ESB sale order line that are actually delivery information.
+
 **Bugfixes**
 
 **Build**